More Animation Trees, Hit Reaction Test Code in place
Easy, Meduim, Hard Zombie Animations Trees have been made. Hit Test Reaction for zombies taking attacks is working in the weapons script but could be in a better place ideally. Just needs balancing when the hit reaction animation is played to not take any damage if the zombie takes damage...